some bug fixes

+++ bin/chandra/xtc_gif_generator.py
@@ -16,6 +16,7 @@ import sys
 import numpy as np
 import astropy.io.fits as pyfits
 import matplotlib.pyplot as plt
+from matplotlib.patches import Circle
 import seaborn as sbs
 from matplotlib.colors import LogNorm
 from astropy.convolution import convolve_fft, convolve, Gaussian2DKernel
@@ -27,6 +28,7 @@ from XtDac.ChandraUtils.run_command import CommandRunner
 from XtDac.ChandraUtils.sanitize_filename import sanitize_filename
 
 from XtDac.DivideAndConquer import XMMWCS
+from XtDac.DivideAndConquer.HardwareUnit import hardwareUnitFactory
 
 if __name__=="__main__":
 
@@ -37,6 +39,7 @@ if __name__=="__main__":
     parser.add_argument("--data_path", help="Path to directory containing data", required=False, type=str, default='.')
 
     parser.add_argument("--verbose-debug", action='store_true')
+    parser.add_argument("--cleanup", action='store_true')
 
     # Get the logger
     logger = logging_system.get_logger(os.path.basename(sys.argv[0]))
@@ -61,29 +64,49 @@ if __name__=="__main__":
 
         duration = tstop-tstart
 
-        event_file = find_files(data_path, "ccd_%s_%s_filtered_nohot.fits" % (ccd, obsid))[0]
+        event_files = find_files(data_path, "ccd_%s_*_filtered_*.fits" % (ccd))
+
+        assert len(event_files)==1, "Couldn't find event file. " \
+                                    "I was looking for %s" % ("ccd_%s_*_filtered_*.fits" % (ccd))
+
+        event_file = event_files[0]
 
         # get start and stop time of observation
-        with pyfits.open(event_file, memmap=False) as event_ext:
+        with pyfits.open(event_file, memmap=False) as fits_file:
 
-            tmin = event_ext['EVENTS'].header['TSTART']
-            tmax = event_ext['EVENTS'].header['TSTOP']
+            # Get the start of the first GTI and the stop of the last one
+            gti_starts = []
+            gti_stops = []
+            for ext in fits_file[1:]:
+
+                if ext.header['EXTNAME'] == 'GTI':
+
+                    gti_starts.append(ext.data.field("START").min())
+                    gti_stops.append(ext.data.field("STOP").max())
+
+            frame_time = fits_file['EVENTS'].header['TIMEDEL']
+
+            tmin = min(gti_starts)
+            tmax = max(gti_stops)
 
             # Get minimum and maximum X and Y, so we use always the same binning for the images
-            xmin, xmax = event_ext['EVENTS'].data.field("X").min(), event_ext['EVENTS'].data.field("X").max()
-            ymin, ymax = event_ext['EVENTS'].data.field("Y").min(), event_ext['EVENTS'].data.field("Y").max()
+            xmin, xmax = fits_file['EVENTS'].data.field("X").min(), fits_file['EVENTS'].data.field("X").max()
+            ymin, ymax = fits_file['EVENTS'].data.field("Y").min(), fits_file['EVENTS'].data.field("Y").max()
 
             # Get position and transform to X,Y
 
             ra = transient['RA']
             dec = transient['Dec']
 
-            wcs = XMMWCS.XMMWCS(event_file, event_ext['EVENTS'].data.field("X"), event_ext['EVENTS'].data.field("Y"))
+            wcs = XMMWCS.XMMWCS(event_file, fits_file['EVENTS'].data.field("X"), fits_file['EVENTS'].data.field("Y"))
             transient_x, transient_y = wcs.sky2xy([[ra, dec]])[0]
+
+        hwu = hardwareUnitFactory(event_file)
 
         logger.info("Duration: %s" %duration)
         logger.info("Tmin: %s" % tmin)
         logger.info("Tmax: %s" %tmax)
+        logger.info("frame time: %s" % frame_time)
         logger.info("Obs Time: %s" %(tmax-tmin))
         logger.info("X,Y = (%.3f, %.3f)" % (transient_x, transient_y))
 
@@ -95,7 +118,7 @@ if __name__=="__main__":
         # Add tstart only if it is sufficiently different from tmin (so we don't get an empty interval when the
         # transient is right at the beginning)
 
-        if abs(tstart - tmin) > 0.1:
+        if abs(tstart - tmin) > (frame_time + frame_time / 2.0):
 
             intervals.append(tstart)
 
@@ -104,9 +127,13 @@ if __name__=="__main__":
         # Add tmax only if it is sufficiently different from tstop, so we don't get an empty interval when the
         # transient is right at the end)
 
-        if abs(tmax - tstop) > 0.1:
+        if abs(tmax - tstop) > (frame_time + frame_time / 2.0):
 
             intervals.append(tmax)
+
+        intervals = sorted(intervals)
+
+        deltas = np.array(intervals[1:]) - np.array(intervals[:-1])
 
         evt_name, evt_file_ext = os.path.splitext(os.path.basename(event_file))
 
@@ -115,7 +142,7 @@ if __name__=="__main__":
 
         for i in range(len(intervals)-1):
 
-            outfile = "%s_TI_%s_cand_%s%s" %(evt_name, i+1, candidate, evt_file_ext)
+            outfile = "cand_%s_%s_TI_%s%s" %(candidate, evt_name, i+1, evt_file_ext)
 
             cmd_line = 'ftcopy \"%s[(TIME >= %s) && (TIME <= %s)]\" %s clobber=yes ' \
                        % (event_file, intervals[i], intervals[i+1], outfile)
@@ -131,19 +158,25 @@ if __name__=="__main__":
         # xbins = np.linspace(xmin, xmax, 300)
         # ybins = np.linspace(ymin, ymax, 300)
 
-        conv_kernel_size = max(float(transient['PSF_sizearcsec']) / 3, 2)
+        conv_kernel_size = max(np.ceil(float(transient['PSF_sizearcsec']) / hwu.getPixelScale()), 5)
 
-        xbins = np.arange(transient_x - 15 * conv_kernel_size, transient_x + 15 * conv_kernel_size, 1)
-        ybins = np.arange(transient_y - 15 * conv_kernel_size, transient_y + 15 * conv_kernel_size, 1)
+        # Create bins of size 1 (i.e., one pixel per bin)
+
+        xbins = np.arange(transient_x - 5 * conv_kernel_size, transient_x + 5 * conv_kernel_size + 1, 1)
+        ybins = np.arange(transient_y - 5 * conv_kernel_size, transient_y + 5 * conv_kernel_size + 1, 1)
 
         logger.info("Image will be %i x %i pixels" % (xbins.shape[0], ybins.shape[0]))
 
         fig = plt.figure()
 
         sub = fig.add_subplot(111)
-        sub.set_title("ObsID %s, CCD %s \nTstart = %s, Tstop = %s" % (obsid, ccd, tstart, tstop))
+        sub.set_title("ObsID %s, CCD %s \nTstart = %s, Tstop = %s (%i events)" % (obsid, ccd, tstart, tstop,
+                                                                                  float(transient['N_events'])))
 
         for i, image in enumerate(images):
+
+            # Compute interval duration
+            dt = intervals[i + 1] - intervals[i]
 
             #get x and y coordinates of image from fits file
             data = pyfits.getdata(image)
@@ -156,7 +189,7 @@ if __name__=="__main__":
             hh, X, Y = np.histogram2d(x, y, bins=[xbins, ybins])
 
             #smooth data
-            gauss_kernel = Gaussian2DKernel(stddev=max(conv_kernel_size / 10.0, 2.0))
+            gauss_kernel = Gaussian2DKernel(stddev=max(conv_kernel_size / 8.0, 2.0))
             smoothed_data_gauss = convolve_fft(hh, gauss_kernel, normalize_kernel=True)
 
             if x.shape[0] > 0:
@@ -168,26 +201,30 @@ if __name__=="__main__":
                 # No events in this image. Generate an empty image
                 img = sub.imshow(smoothed_data_gauss, cmap='hot', animated=True, origin='lower')
 
-            text = sub.annotate("%i / %i" % (i+1, len(images)), xy=(0.5, 0.03),
-                                xycoords='figure fraction', annotation_clip=False)
+            # Draw PSF circle
+            circ = Circle((smoothed_data_gauss.shape[0]/2.0 + 1, smoothed_data_gauss.shape[1]/2.0 + 1),
+                          float(transient['PSF_sizearcsec']) / hwu.getPixelScale(),
+                          fill=False, lw=2, color='green')
+            sub.add_patch(circ)
 
-            # Compute interval duration
-            dt = intervals[i+1] - intervals[i]
+            text = sub.annotate("%i" % (i+1), xy=(0.5, 0.03),
+                                xycoords='figure fraction',
+                                annotation_clip=False)
 
-            text2 = sub.annotate("%i events" % (float(transient['N_events'])), xy=(0.03, 0.45),
-                                xycoords='figure fraction', annotation_clip=False)
-
-            text3 = sub.annotate("Duration: %.2f s" % (dt), xy=(0.03, 0.55),
-                                 xycoords='figure fraction', annotation_clip=False)
+            text3 = sub.annotate("Duration: %.2f s" % (dt), xy=(0.55, 0.13),
+                                 xycoords='figure fraction',
+                                 annotation_clip=False, color='green')
 
             sub.set_yticklabels([])
             sub.set_xticklabels([])
 
             #add this image to list of frames
-            frames.append([img, text, text2, text3])
+            frames.append([img, text, text3, circ])
 
             # Remove the image
-            os.remove(image)
+            if args.cleanup:
+
+                os.remove(image)
 
         #animate and save gif
         logger.info("Creating gif ObsID %s, CCD %s, Candidate %s...\n" %(obsid, ccd, candidate))
